Remove custom arm height :(
There was an issue where selecting an item from being empty handed in 3rd person, then switching back to 1st person caused the animation controller to get stuck and the arm to appear fully "down". I couldn't figure out a fix for this issue, and even if I did, there might still be hidden bugs that I wouldn't be aware of. `v.player_arm_height` used to not have the issue that this code attempted to fix. Ideally, Mojang should fix it, but we all know that's never going to happen